[miant] update binder to use conda nightly and jupyter-remote-desktop-proxy (#7352)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7331

# Description
This PR does 2 main things
1. updates the apt requirements and switches to the up-to-date
`jupyter-remote-desktop-proxy`
2. switches to using the conda nightly channel to install a napari
nightly, rather than the extra pip step and specifying all the
requirements. This is quite a bit faster to spin up too.
I mean we could just go with a release, but there is something to a
`main` install for someone to test something?

You can see it in action in my fork/branch:

https://mybinder.org/v2/gh/psobolewskiPhD/napari/fix_binder?urlpath=desktop

Only small thing is there is no feedback when you click the napari icon,
so easy to launch 2, but I think that's just how this linux window
system works.